Fibonacci

+++ Week7/Day1/Fibonacci/src/Fibonacci.java
@@ -2,23 +2,23 @@ import java.util.ArrayList;
 
 public class Fibonacci {
 
-    ArrayList<Integer> fibonacciNrs;
+    ArrayList<Long> fibonacciNrs;
 
     public Fibonacci() {
         this.fibonacciNrs = new ArrayList<>();
     }
 
-    public int fibonacci(int index) {
+    public long fibonacci(int index) {
 
 
-        fibonacciNrs.add(0);
-        fibonacciNrs.add(1);
+        fibonacciNrs.add((long)0);
+        fibonacciNrs.add((long)1);
 
         int k = 2;
 
         while (fibonacciNrs.size() <= (index+1) ) {
 
-            int sumOfPreviousTwo = fibonacciNrs.get(k - 2) + fibonacciNrs.get(k - 1);
+            Long sumOfPreviousTwo = Long.valueOf(fibonacciNrs.get(k - 2) + fibonacciNrs.get(k - 1));
             fibonacciNrs.add(sumOfPreviousTwo);
             k++;
